Code refresh

+++ AlgorithmsStrings - CamelCase/AlgorithmsStrings - CamelCase.cpp
@@ -9,8 +9,10 @@ int main()
 	std::string str{}; std::getline(std::cin, str);
 	const std::size_t count = std::count_if(std::begin(str), std::end(str),
 		[](const char charector) noexcept -> bool {
-			return std::isupper(charector);
-		});
+			return std::isupper(static_cast<unsigned int>(charector));
+		}
+	);
 	std::cout << count + 1 << '\n';
 	return 0;
 }
+
